Optimization & Update Query

+++ src/modules/record/record.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  BadRequestException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Storage } from '@google-cloud/storage';
@@ -17,12 +21,12 @@ export class RecordService {
   ) {}
 
   async create(createRecordDto: CreateRecordDto): Promise<Record> {
-    const record = this.recordRepository.create({
-      ...createRecordDto,
-      user: { id: createRecordDto.id_user } as User,
-    });
-
-    return this.recordRepository.save(record);
+    return this.recordRepository.save(
+      this.recordRepository.create({
+        ...createRecordDto,
+        user: { id: createRecordDto.id_user } as User,
+      }),
+    );
   }
 
   async findAll(): Promise<Record[]> {
@@ -32,7 +36,7 @@ export class RecordService {
   }
 
   async findAllByUser(id_user: number): Promise<Record[]> {
-    return await this.recordRepository.find({
+    return this.recordRepository.find({
       where: { user: { id: id_user } },
       relations: ['diseaseRecords', 'diseaseRecords.disease'],
     });
@@ -40,22 +44,28 @@ export class RecordService {
 
   async findOne(id: number): Promise<Record> {
     const record = await this.recordRepository.findOne({
-      where: { id: +id },
-      relations: ['diseaseRecords', 'diseaseRecords.disease'],
+      where: { id },
+      relations: ['diseaseRecords', 'diseaseRecords.disease', 'user'],
     });
+
     if (!record) throw new NotFoundException(`Record with id ${id} not found`);
+    if (record.user) delete record.user.password;
+
     return record;
   }
 
-  async remove(id: number): Promise<{ message: string }> {
-    try {
-      const record = await this.findOne(id);
-      if (record.image) await this.deleteFromGCS(record.image);
-      await this.recordRepository.remove(record);
-      return { message: `Record with id ${id} successfully deleted.` };
-    } catch (err) {
-      throw new NotFoundException(`Failed to delete record: ${err.message}`);
-    }
+  async remove(id: number, id_user: number): Promise<{ message: string }> {
+    const record = await this.findOne(id);
+
+    if (record.user.id != id_user)
+      throw new BadRequestException(
+        'User does not have permission to delete this record.',
+      );
+    if (record.image) await this.deleteFromGCS(record.image);
+
+    await this.recordRepository.remove(record);
+
+    return { message: `Record with id ${id} successfully deleted.` };
   }
 
   private async deleteFromGCS(imageUrl: string): Promise<void> {
@@ -65,8 +75,7 @@ export class RecordService {
     try {
       await this.storageClient.bucket(this.bucketName).file(filePath).delete();
     } catch (err) {
-      if (err.code === 404)
-        throw new NotFoundException('File not found');
+      if (err.code === 404) throw new NotFoundException('File not found');
       throw new Error(`Failed to delete file: ${err.message}`);
     }
   }
